refactor: modify stream practice - move StreamExercise to StreamPractice under java-practice
